speedups

+++ ProteoformSuiteInternal/ProteoformRelation.cs
@@ -86,7 +86,7 @@ namespace ProteoformSuiteInternal
                 && pf2 as TheoreticalProteoform != null)
             {
                 TheoreticalProteoform t = pf2 as TheoreticalProteoform;
-                double mass_tolerance = t.modified_mass / 1000000 * (double)Sweet.lollipop.mass_tolerance;
+                double mass_tolerance = t.modified_mass / 1000000 * Sweet.lollipop.mass_tolerance;
                 List<PtmSet> narrower_range_of_candidates = candidate_sets.Where(s => Math.Abs(s.mass - delta_mass) < 0.05).ToList();
                 candidate_ptmset = t.generate_possible_added_ptmsets(narrower_range_of_candidates, delta_mass, mass_tolerance, Sweet.lollipop.theoretical_database.all_mods_with_mass, t, t.sequence, Sweet.lollipop.mod_rank_first_quartile)
                     .OrderBy(x => x.ptm_rank_sum + Math.Abs(Math.Abs(x.mass) - Math.Abs(delta_mass)) * 10E-6) // major score: delta rank; tie breaker: deltaM, where it's always less than 1
